API

Note

We try to document the basics here, but it is not meant as a replacement for the upstream documentation. If you need general help we recommend looking at the test engine wiki, and possibly at the documentation comments in these header files:

imgui_te_engine.h (for the engine API)

imgui_te_context.h (for the test context API)

There are two major parts of the test engine:

The Engine itself. This is the class that executes the tests and handles things like interacting with the GUI.

The Test context API, which is what you'll use to control the GUI and write the tests.

Danger

For the sake of simplicitly certain parts of the API are not memory-safe. This means that some test engine types are wrapped as raw pointers that are owned by C++ rather than Julia, which means that using them after they have been free'd will cause segfaults. All memory-unsafe types are marked as such in their docstrings.

Because of all that, we recommend using such types only temporarily in the style recommended by the upstream examples. This style is good:

# The test object is never even assigned to a variable
@register_test(engine, "foo", "bar") do ctx
    ...
end

This style is less good:

all_tests = []
t = @register_test(engine, "foo", "bar")
t.TestFunc = ...

# Dangerous because it allows `t` to potentially be accessed after the
# engine has been destroyed.
push!(all_tests, t)

Note that in all the examples in the docstrings below we assume that we have already evaluated:

import CImGui as ig
using ImGuiTestEngine
import ImGuiTestEngine as te

API

Engine

Registering tests

Test context

Other

ImGuiTestRef

ImGuiTestRefDesc

ImGuiTestGenericItemStatus

ImGuiTestItemList

ImGuiPerfTool

ImGuiCaptureContext

ImGuiCaptureToolUI

ImGuiCsvParser

ImGuiTestLog

Engine

ImGuiTestEngine.Engine — Type

mutable struct Engine

Represents a test engine context. This a wrapper around the upstream ImGuiTestEngine type. Don't create it yourself, use CreateContext().

ImGuiTestEngine.CreateContext — Function

CreateContext(
;
    exit_on_completion,
    show_test_window
) -> ImGuiTestEngine.Engine

Create a test engine context. The keyword arguments don't do anything in this library, they're used to support the test engine in CImGui.jl's renderloop.

Arguments

exit_on_completion=true: Exit the program after the tests have completed.

show_test_window=true: Call ShowTestEngineWindows() while running the tests.

Examples

engine = te.CreateContext()

ImGuiTestEngine.DestroyContext — Function

DestroyContext(engine::ImGuiTestEngine.Engine; throw)

Destroy a test engine context.

Arguments

throw=true: Whether to throw an exception if the engine has already been destroyed.

Examples

engine = te.CreateContext()
te.DestroyContext(engine)

ImGuiTestEngine.ShowTestEngineWindows — Function

ShowTestEngineWindows(
    engine::ImGuiTestEngine.Engine,
    p_open
)

Upstream link.

ImGuiTestEngine.OpenSourceFile — Function

OpenSourceFile(
    engine::ImGuiTestEngine.Engine,
    source_filename,
    source_line_no
)

Upstream link.

ImGuiTestEngine.PrintResultSummary — Function

PrintResultSummary(engine::ImGuiTestEngine.Engine)

Upstream link.

ImGuiTestEngine.Export — Function

Export(engine::ImGuiTestEngine.Engine)

Upstream link.

ImGuiTestEngine.ExportEx — Function

ExportEx(
    engine::ImGuiTestEngine.Engine,
    format::ImGuiTestEngine.lib.ImGuiTestEngineExportFormat,
    filename
)

Upstream link.

ImGuiTestEngine.Start — Function

Start(
    engine::ImGuiTestEngine.Engine,
    ui_ctx::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiContext}}
)

Bind to a dear imgui context. Start coroutine.

Upstream link.

ImGuiTestEngine.Stop — Function

Stop(engine::ImGuiTestEngine.Engine)

Stop coroutine and export if any. (Unbind will lazily happen on context shutdown).

Upstream link.

ImGuiTestEngine.PostSwap — Function

PostSwap(engine::ImGuiTestEngine.Engine)

Call every frame after framebuffer swap, will process screen capture and call test_io.ScreenCaptureFunc().

Upstream link.

ImGuiTestEngine.GetIO — Function

GetIO(
    engine::ImGuiTestEngine.Engine
) -> Ptr{ImGuiTestEngine.lib.ImGuiTestEngineIO}

Upstream link.

ImGuiTestEngine.RegisterTest — Function

RegisterTest(
    engine::ImGuiTestEngine.Engine,
    category,
    name
) -> Ptr{ImGuiTestEngine.lib.ImGuiTest}
RegisterTest(
    engine::ImGuiTestEngine.Engine,
    category,
    name,
    src_file
) -> Ptr{ImGuiTestEngine.lib.ImGuiTest}
RegisterTest(
    engine::ImGuiTestEngine.Engine,
    category,
    name,
    src_file,
    src_line
) -> Ptr{ImGuiTestEngine.lib.ImGuiTest}

Prefer calling IMREGISTERTEST().

Upstream link.

ImGuiTestEngine.UnregisterTest — Function

UnregisterTest(
    engine::ImGuiTestEngine.Engine,
    test::ImGuiTestEngine.ImGuiTest
)

Upstream link.

ImGuiTestEngine.UnregisterAllTests — Function

UnregisterAllTests(engine::ImGuiTestEngine.Engine)

Upstream link.

ImGuiTestEngine.QueueTest — Function

QueueTest(
    engine::ImGuiTestEngine.Engine,
    test::ImGuiTestEngine.ImGuiTest
)
QueueTest(
    engine::ImGuiTestEngine.Engine,
    test::ImGuiTestEngine.ImGuiTest,
    run_flags
)

Upstream link.

ImGuiTestEngine.QueueTests — Function

QueueTests(
    engine::ImGuiTestEngine.Engine,
    group::ImGuiTestEngine.lib.ImGuiTestGroup
)
QueueTests(
    engine::ImGuiTestEngine.Engine,
    group::ImGuiTestEngine.lib.ImGuiTestGroup,
    filter
)
QueueTests(
    engine::ImGuiTestEngine.Engine,
    group::ImGuiTestEngine.lib.ImGuiTestGroup,
    filter,
    run_flags
)

Upstream link.

ImGuiTestEngine.TryAbortEngine — Function

TryAbortEngine(engine::ImGuiTestEngine.Engine) -> Bool

Upstream link.

ImGuiTestEngine.AbortCurrentTest — Function

AbortCurrentTest(engine::ImGuiTestEngine.Engine)

Upstream link.

ImGuiTestEngine.FindTestByName — Function

FindTestByName(
    engine::ImGuiTestEngine.Engine,
    category,
    name
) -> Ptr{ImGuiTestEngine.lib.ImGuiTest}

Upstream link.

ImGuiTestEngine.IsTestQueueEmpty — Function

IsTestQueueEmpty(engine::ImGuiTestEngine.Engine) -> Bool

Upstream link.

ImGuiTestEngine.IsUsingSimulatedInputs — Function

IsUsingSimulatedInputs(
    engine::ImGuiTestEngine.Engine
) -> Bool

Upstream link.

ImGuiTestEngine.GetResultSummary — Function

GetResultSummary(
    engine::ImGuiTestEngine.Engine,
    out_results::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestEngineResultSummary}}
)

Upstream link.

ImGuiTestEngine.GetTestList — Function

GetTestList(
    engine::ImGuiTestEngine.Engine,
    out_tests::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImVector_ImGuiTest_Ptr}}
)

Upstream link.

ImGuiTestEngine.GetTestQueue — Function

GetTestQueue(
    engine::ImGuiTestEngine.Engine,
    out_tests::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImVector_ImGuiTestRunTask}}
)

Upstream link.

Registering tests

Once the engine is set up you can register some tests for it to run:

ImGuiTestEngine.@register_test — Macro

@register_test(engine, category::AbstractString, name::AbstractString)::ImGuiTest
@register_test(f::Function, engine,
               category::AbstractString, name::AbstractString)::ImGuiTest

Register a ImGuiTest. Note that it will not be executed until the test is queued, either programmatically with QueueTests() or by the user running it manually through ShowTestEngineWindows().

Examples

If you only need to set TestFunc you can use do-syntax:

engine = te.CreateContext()
@register_test(engine, "foo", "bar") do
    @imtest ctx isa te.TestContext
end

To set GuiFunc as well you'll need to set the GuiFunc property:

engine = te.CreateContext()
t = @register_test(engine, "foo", "bar")
t.GuiFunc = () -> begin
    ig.Begin("Foo")
    ig.End()
end
t.TestFunc = () -> @info "Hello world!"

ImGuiTestEngine.ImGuiTest — Type

mutable struct ImGuiTest

Wrapper around the upstream ImGuiTest. Don't create this yourself, use @register_test(). Once it's created you can assign functions to these properties:

GuiFunc::Function, for standalone GUI code that you want to run/test. This shouldn't be necessary if you're testing your own GUI.

TestFunc::Function, for tests that you want to execute.

The functions you assign should not take any arguments, and any return value will be discarded.

Danger

This a memory-unsafe type, only use it while the engine is alive.

ImGuiTestEngine.SetOwnedName — Function

SetOwnedName(self::Ptr{ImGuiTestEngine.ImGuiTest}, name)

Upstream link.

Within the tests you will often want to refer to parts of your interface by named reference. In the C++ API this is done with the ImGuiTestRef type but with ImGuiTestEngine.jl you should use either strings or integers and they will automatically be converted.

Test context

Inside GuiFunc and TestFunc you can use any methods of the test context API to control and test the GUI. It's not safe to use them outside of a GuiFunc/TestFunc.

ImGuiTestEngine.@imcheck — Macro

@imcheck expr

A port of the upstream IM_CHECK() macro. Like the upstream macro, this will return early from the calling function if expr evaluates to false. Prefer using it over @test because it will register test results with the test engine, which can be convenient if you're using the built-in test engine window (see ShowTestEngineWindows()).

@imcheck hooks into @testset's by default, so a failure will be recorded with your Julia Test tests as well as with the test engine. If this is not wanted it can be disabled by passing jltest=false.

Note

A limitation of the current implementation is that nicely parsing the expression, e.g. to display both arguments of an equality, is not supported.

Examples

engine = te.CreateContext()
@register_test(engine, "foo", "bar") do
    # This record the result with `Test` as well as the test engine
    @imcheck false

    # This will only record the result with the test engine
    @imcheck false jltest=false
end

ImGuiTestEngine.@imcheck_noret — Macro

@imcheck_noret expr

Same as @imcheck, except that it will not return early from the calling function.

ImGuiTestEngine.OpenAndClose — Function

OpenAndClose(
    f,
    test_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

A helper function that will ensure test_ref is open, execute f(), and close test_ref again. A typical use would be to open a section, run some tests, and then close the section again (handy for re-runnable tests).

Examples

@register_test(engine, "foo", "bar") do
    OpenAndClose("My section") do
        # ...
    end
end

OpenAndClose(
    test_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Open and then close test_ref.

Examples

@register_test(engine, "foo", "bar") do
    OpenAndClose("My section")
end

ImGuiTestEngine.Finish — Function

Finish()
Finish(status::ImGuiTestEngine.lib.ImGuiTestStatus)

Set test status and stop running. Usually called when running test logic from GuiFunc() only.

Upstream link.

ImGuiTestEngine.RunChildTest — Function

RunChildTest(
    test_name
) -> ImGuiTestEngine.lib.ImGuiTestStatus
RunChildTest(
    test_name,
    flags
) -> ImGuiTestEngine.lib.ImGuiTestStatus

[Experimental] Run another test from the current test.

Upstream link.

ImGuiTestEngine.IsError — Function

IsError() -> Bool

Upstream link.

ImGuiTestEngine.IsWarmUpGuiFrame — Function

IsWarmUpGuiFrame() -> Bool

Unless test->Flags has ImGuiTestFlags_NoGuiWarmUp, we run GuiFunc() twice before running TestFunc(). Those frames are called "WarmUp" frames.

Upstream link.

ImGuiTestEngine.IsFirstGuiFrame — Function

IsFirstGuiFrame() -> Bool

Upstream link.

ImGuiTestEngine.IsFirstTestFrame — Function

IsFirstTestFrame() -> Bool

First frame where TestFunc is running (after warm-up frame).

Upstream link.

ImGuiTestEngine.IsGuiFuncOnly — Function

IsGuiFuncOnly() -> Bool

Upstream link.

ImGuiTestEngine.SuspendTestFunc — Function

SuspendTestFunc() -> Bool
SuspendTestFunc(file) -> Bool
SuspendTestFunc(file, line) -> Bool

[DEBUG] Generally called via IMSUSPENDTESTFUNC.

Upstream link.

ImGuiTestEngine.LogEx — Function

LogEx(
    level::ImGuiTestEngine.lib.ImGuiTestVerboseLevel,
    flags,
    fmt
)

Upstream link.

ImGuiTestEngine.LogToTTY — Function

LogToTTY(
    level::ImGuiTestEngine.lib.ImGuiTestVerboseLevel,
    message
)
LogToTTY(
    level::ImGuiTestEngine.lib.ImGuiTestVerboseLevel,
    message,
    message_end
)

Upstream link.

ImGuiTestEngine.LogToDebugger — Function

LogToDebugger(
    level::ImGuiTestEngine.lib.ImGuiTestVerboseLevel,
    message
)

Upstream link.

ImGuiTestEngine.LogDebug — Function

LogDebug(fmt)

ImGuiTestVerboseLevelDebug or ImGuiTestVerboseLevelTrace depending on context depth.

Upstream link.

ImGuiTestEngine.LogInfo — Function

LogInfo(fmt)

ImGuiTestVerboseLevel_Info.

Upstream link.

ImGuiTestEngine.LogWarning — Function

LogWarning(fmt)

ImGuiTestVerboseLevel_Warning.

Upstream link.

ImGuiTestEngine.LogError — Function

LogError(fmt)

ImGuiTestVerboseLevel_Error.

Upstream link.

ImGuiTestEngine.LogBasicUiState — Function

LogBasicUiState()

Upstream link.

ImGuiTestEngine.LogItemList — Function

LogItemList(
    list::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestItemList}}
)

Upstream link.

ImGuiTestEngine.Yield — Function

Yield(engine::ImGuiTestEngine.Engine)

Warning

This function is internal, it may change in the future.

Upstream link.

Yield()
Yield(count)

Upstream link.

ImGuiTestEngine.Sleep — Function

Sleep(time_in_second)

Sleep for a given simulation time, unless in Fast mode.

Upstream link.

ImGuiTestEngine.SleepShort — Function

SleepShort()

Standard short delay of io.ActionDelayShort (~0.15f), unless in Fast mode.

Upstream link.

ImGuiTestEngine.SleepStandard — Function

SleepStandard()

Standard regular delay of io.ActionDelayStandard (~0.40f), unless in Fast mode.

Upstream link.

ImGuiTestEngine.SleepNoSkip — Function

SleepNoSkip(time_in_second, framestep_in_second)

Upstream link.

ImGuiTestEngine.SetRef — Function

SetRef(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

SetRef(window::Ref{CImGui.lib.ImGuiWindow})

Shortcut to SetRef(window->Name) which works for ChildWindow (see code).

Upstream link.

ImGuiTestEngine.GetRef — Function

GetRef() -> ImGuiTestEngine.lib.ImGuiTestRef

Upstream link.

ImGuiTestEngine.WindowInfo — Function

WindowInfo(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> ImGuiTestEngine.lib.ImGuiTestItemInfo
WindowInfo(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
) -> ImGuiTestEngine.lib.ImGuiTestItemInfo

Upstream link.

ImGuiTestEngine.WindowClose — Function

WindowClose(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.WindowCollapse — Function

WindowCollapse(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    collapsed
)

Upstream link.

ImGuiTestEngine.WindowFocus — Function

WindowFocus(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
WindowFocus(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.WindowBringToFront — Function

WindowBringToFront(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
WindowBringToFront(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.WindowMove — Function

WindowMove(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    pos::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
)
WindowMove(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    pos::Union{CImGui.lib.ImVec2, Tuple{T, T} where T},
    pivot::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
)
WindowMove(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    pos::Union{CImGui.lib.ImVec2, Tuple{T, T} where T},
    pivot::Union{CImGui.lib.ImVec2, Tuple{T, T} where T},
    flags
)

Upstream link.

ImGuiTestEngine.WindowResize — Function

WindowResize(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    sz::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
)

Upstream link.

ImGuiTestEngine.WindowTeleportToMakePosVisible — Function

WindowTeleportToMakePosVisible(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    pos_in_window::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
) -> Bool

Upstream link.

ImGuiTestEngine.GetWindowByRef — Function

GetWindowByRef(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Ptr{CImGui.lib.ImGuiWindow}

Upstream link.

ImGuiTestEngine.PopupCloseOne — Function

PopupCloseOne()

Upstream link.

ImGuiTestEngine.PopupCloseAll — Function

PopupCloseAll()

Upstream link.

ImGuiTestEngine.PopupGetWindowID — Function

PopupGetWindowID(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> UInt32

Upstream link.

ImGuiTestEngine.GetID — Function

GetID(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> UInt32

Upstream link.

GetID(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    seed_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> UInt32

Upstream link.

ImGuiTestEngine.GetPosOnVoid — Function

GetPosOnVoid(
    viewport::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiViewport}}
) -> Any

Find a point that has no windows // FIXME: This needs error return and flag to enable/disable forcefully finding void.

Upstream link.

ImGuiTestEngine.GetWindowTitlebarPoint — Function

GetWindowTitlebarPoint(
    window_ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Any

Return a clickable point on window title-bar (window tab for docked windows).

Upstream link.

ImGuiTestEngine.GetMainMonitorWorkPos — Function

GetMainMonitorWorkPos() -> Any

Work pos and size of main viewport when viewports are disabled, or work pos and size of monitor containing main viewport when viewports are enabled.

Upstream link.

ImGuiTestEngine.GetMainMonitorWorkSize — Function

GetMainMonitorWorkSize() -> Any

Upstream link.

ImGuiTestEngine.CaptureReset — Function

CaptureReset()

Reset state (use when doing multiple captures).

Upstream link.

ImGuiTestEngine.CaptureSetExtension — Function

CaptureSetExtension(ext)

Set capture file format (otherwise for video this default to EngineIO->VideoCaptureExtension).

Upstream link.

ImGuiTestEngine.CaptureAddWindow — Function

CaptureAddWindow(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Bool

Add window to be captured (default to capture everything).

Upstream link.

ImGuiTestEngine.CaptureScreenshotWindow — Function

CaptureScreenshotWindow(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
CaptureScreenshotWindow(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    capture_flags
)

Trigger a screen capture of a single window (== CaptureAddWindow() + CaptureScreenshot()).

Upstream link.

ImGuiTestEngine.CaptureScreenshot — Function

CaptureScreenshot(
    engine::ImGuiTestEngine.Engine,
    args::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiCaptureArgs}}
) -> Bool

Warning

This function is internal, it may change in the future.

Upstream link.

CaptureScreenshot() -> Bool
CaptureScreenshot(capture_flags) -> Bool

Trigger a screen capture.

Upstream link.

ImGuiTestEngine.CaptureBeginVideo — Function

CaptureBeginVideo(
    engine::ImGuiTestEngine.Engine,
    args::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiCaptureArgs}}
) -> Bool

Warning

This function is internal, it may change in the future.

Upstream link.

CaptureBeginVideo() -> Bool

Start a video capture.

Upstream link.

ImGuiTestEngine.CaptureEndVideo — Function

CaptureEndVideo(
    engine::ImGuiTestEngine.Engine,
    args::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiCaptureArgs}}
) -> Bool

Warning

This function is internal, it may change in the future.

Upstream link.

CaptureEndVideo() -> Bool

Upstream link.

ImGuiTestEngine.MouseMove — Function

MouseMove(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
MouseMove(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.MouseMoveToPos — Function

MouseMoveToPos(
    pos::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
)

Upstream link.

ImGuiTestEngine.MouseTeleportToPos — Function

MouseTeleportToPos(
    pos::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
)
MouseTeleportToPos(
    pos::Union{CImGui.lib.ImVec2, Tuple{T, T} where T},
    flags
)

Upstream link.

ImGuiTestEngine.MouseClick — Function

MouseClick()
MouseClick(button)

Upstream link.

ImGuiTestEngine.MouseClickMulti — Function

MouseClickMulti(button, count)

Upstream link.

ImGuiTestEngine.MouseDoubleClick — Function

MouseDoubleClick()
MouseDoubleClick(button)

Upstream link.

ImGuiTestEngine.MouseDown — Function

MouseDown()
MouseDown(button)

Upstream link.

ImGuiTestEngine.MouseUp — Function

MouseUp()
MouseUp(button)

Upstream link.

ImGuiTestEngine.MouseLiftDragThreshold — Function

MouseLiftDragThreshold()
MouseLiftDragThreshold(button)

Upstream link.

ImGuiTestEngine.MouseDragWithDelta — Function

MouseDragWithDelta(
    delta::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
)
MouseDragWithDelta(
    delta::Union{CImGui.lib.ImVec2, Tuple{T, T} where T},
    button
)

Upstream link.

ImGuiTestEngine.MouseWheel — Function

MouseWheel(
    delta::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
)

Upstream link.

ImGuiTestEngine.MouseWheelX — Function

MouseWheelX(dx)

Upstream link.

ImGuiTestEngine.MouseWheelY — Function

MouseWheelY(dy)

Upstream link.

ImGuiTestEngine.MouseMoveToVoid — Function

MouseMoveToVoid()
MouseMoveToVoid(
    viewport::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiViewport}}
)

Upstream link.

ImGuiTestEngine.MouseClickOnVoid — Function

MouseClickOnVoid()
MouseClickOnVoid(button)
MouseClickOnVoid(
    button,
    viewport::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiViewport}}
)

Upstream link.

ImGuiTestEngine.FindHoveredWindowAtPos — Function

FindHoveredWindowAtPos(
    pos::Union{Ptr{Nothing}, Ref{CImGui.lib.ImVec2}, Ref{Tuple{T, T} where T}}
) -> Ptr{CImGui.lib.ImGuiWindow}

Upstream link.

ImGuiTestEngine.FindExistingVoidPosOnViewport — Function

FindExistingVoidPosOnViewport(
    viewport::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiViewport}},
    out::Union{Ptr{Nothing}, Ref{CImGui.lib.ImVec2}, Ref{Tuple{T, T} where T}}
) -> Bool

Upstream link.

ImGuiTestEngine.MouseSetViewport — Function

MouseSetViewport(
    window::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiWindow}}
)

Upstream link.

ImGuiTestEngine.MouseSetViewportID — Function

MouseSetViewportID(viewport_id)

Upstream link.

ImGuiTestEngine.KeyDown — Function

KeyDown(key_chord)

Upstream link.

ImGuiTestEngine.KeyUp — Function

KeyUp(key_chord)

Upstream link.

ImGuiTestEngine.KeyPress — Function

KeyPress(key_chord)
KeyPress(key_chord, count)

Upstream link.

ImGuiTestEngine.KeyHold — Function

KeyHold(key_chord, time)

Upstream link.

ImGuiTestEngine.KeySetEx — Function

KeySetEx(key_chord, is_down, time)

Upstream link.

ImGuiTestEngine.KeyChars — Function

KeyChars(chars)

Input characters.

Upstream link.

ImGuiTestEngine.KeyCharsAppend — Function

KeyCharsAppend(chars)

Input characters at end of field.

Upstream link.

ImGuiTestEngine.KeyCharsAppendEnter — Function

KeyCharsAppendEnter(chars)

Input characters at end of field, press Enter.

Upstream link.

ImGuiTestEngine.KeyCharsReplace — Function

KeyCharsReplace(chars)

Delete existing field then input characters.

Upstream link.

ImGuiTestEngine.KeyCharsReplaceEnter — Function

KeyCharsReplaceEnter(chars)

Delete existing field then input characters, press Enter.

Upstream link.

ImGuiTestEngine.SetInputMode — Function

SetInputMode(input_mode::CImGui.lib.ImGuiInputSource)

Mouse or Keyboard or Gamepad. In Keyboard or Gamepad mode, actions such as ItemClick or ItemInput are using nav facilities instead of Mouse.

Upstream link.

ImGuiTestEngine.NavMoveTo — Function

NavMoveTo(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.NavActivate — Function

NavActivate()

Activate current selected item: activate button, tweak sliders/drags. Equivalent of pressing Space on keyboard, ImGuiKey_GamepadFaceUp on a gamepad.

Upstream link.

ImGuiTestEngine.NavInput — Function

NavInput()

Input into select item: input sliders/drags. Equivalent of pressing Enter on keyboard, ImGuiKey_GamepadFaceDown on a gamepad.

Upstream link.

ImGuiTestEngine.ScrollTo — Function

ScrollTo(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    axis::CImGui.lib.ImGuiAxis,
    scroll_v
)
ScrollTo(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    axis::CImGui.lib.ImGuiAxis,
    scroll_v,
    flags
)

Upstream link.

ImGuiTestEngine.ScrollToX — Function

ScrollToX(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    scroll_x
)

Upstream link.

ImGuiTestEngine.ScrollToY — Function

ScrollToY(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    scroll_y
)

Upstream link.

ImGuiTestEngine.ScrollToTop — Function

ScrollToTop(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.ScrollToBottom — Function

ScrollToBottom(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.ScrollToItem — Function

ScrollToItem(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    axis::CImGui.lib.ImGuiAxis
)
ScrollToItem(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    axis::CImGui.lib.ImGuiAxis,
    flags
)

Upstream link.

ImGuiTestEngine.ScrollToItemX — Function

ScrollToItemX(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.ScrollToItemY — Function

ScrollToItemY(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.ScrollToTabItem — Function

ScrollToTabItem(
    tab_bar::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiTabBar}},
    tab_id
)

Upstream link.

ImGuiTestEngine.ScrollErrorCheck — Function

ScrollErrorCheck(
    axis::CImGui.lib.ImGuiAxis,
    expected,
    actual,
    remaining_attempts
) -> Bool

Upstream link.

ImGuiTestEngine.ScrollVerifyScrollMax — Function

ScrollVerifyScrollMax(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.ItemInfo — Function

ItemInfo(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> ImGuiTestEngine.lib.ImGuiTestItemInfo
ItemInfo(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
) -> ImGuiTestEngine.lib.ImGuiTestItemInfo

Upstream link.

ImGuiTestEngine.ItemInfoOpenFullPath — Function

ItemInfoOpenFullPath(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> ImGuiTestEngine.lib.ImGuiTestItemInfo
ItemInfoOpenFullPath(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
) -> ImGuiTestEngine.lib.ImGuiTestItemInfo

Upstream link.

ImGuiTestEngine.ItemInfoHandleWildcardSearch — Function

ItemInfoHandleWildcardSearch(
    wildcard_prefix_start,
    wildcard_prefix_end,
    wildcard_suffix_start
) -> UInt32

Upstream link.

ImGuiTestEngine.ItemInfoNull — Function

ItemInfoNull() -> ImGuiTestEngine.lib.ImGuiTestItemInfo

Upstream link.

ImGuiTestEngine.GatherItems — Function

GatherItems(
    out_list::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestItemList}},
    parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
GatherItems(
    out_list::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestItemList}},
    parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    depth
)

Upstream link.

ImGuiTestEngine.ItemAction — Function

ItemAction(
    action::ImGuiTestEngine.lib.ImGuiTestAction,
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemAction(
    action::ImGuiTestEngine.lib.ImGuiTestAction,
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)
ItemAction(
    action::ImGuiTestEngine.lib.ImGuiTestAction,
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags,
    action_arg
)

Upstream link.

ImGuiTestEngine.ItemClick — Function

ItemClick(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemClick(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    button
)
ItemClick(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    button,
    flags
)

Upstream link.

ImGuiTestEngine.ItemDoubleClick — Function

ItemDoubleClick(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemDoubleClick(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.ItemCheck — Function

ItemCheck(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemCheck(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.ItemUncheck — Function

ItemUncheck(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemUncheck(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.ItemOpen — Function

ItemOpen(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemOpen(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.ItemClose — Function

ItemClose(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemClose(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.ItemInput — Function

ItemInput(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemInput(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.ItemNavActivate — Function

ItemNavActivate(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemNavActivate(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    flags
)

Upstream link.

ImGuiTestEngine.ItemActionAll — Function

ItemActionAll(
    action::ImGuiTestEngine.lib.ImGuiTestAction,
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemActionAll(
    action::ImGuiTestEngine.lib.ImGuiTestAction,
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    filter::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestActionFilter}}
)

Upstream link.

ImGuiTestEngine.ItemOpenAll — Function

ItemOpenAll(
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemOpenAll(
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    depth
)
ItemOpenAll(
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    depth,
    passes
)

Upstream link.

ImGuiTestEngine.ItemCloseAll — Function

ItemCloseAll(
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemCloseAll(
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    depth
)
ItemCloseAll(
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    depth,
    passes
)

Upstream link.

ImGuiTestEngine.ItemInputValue — Function

ItemInputValue(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    v::Integer
)

Upstream link.

ItemInputValue(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    f::Real
)

Upstream link.

ItemInputValue(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    str::Union{Ptr{Int8}, Ptr{Nothing}, String}
)

Upstream link.

ImGuiTestEngine.ItemReadAsInt — Function

ItemReadAsInt(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Int32

Upstream link.

ImGuiTestEngine.ItemReadAsFloat — Function

ItemReadAsFloat(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Float32

Upstream link.

ImGuiTestEngine.ItemReadAsScalar — Function

ItemReadAsScalar(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    data_type,
    out_data
) -> Bool
ItemReadAsScalar(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    data_type,
    out_data,
    flags
) -> Bool

Upstream link.

ImGuiTestEngine.ItemReadAsString — Function

ItemReadAsString(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Ptr{Int8}

Upstream link.

ItemReadAsString(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    out_buf::Union{Ptr{Int8}, Ptr{Nothing}, String},
    out_buf_size::Real
) -> UInt64

Upstream link.

ImGuiTestEngine.ItemExists — Function

ItemExists(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Bool

Upstream link.

ImGuiTestEngine.ItemIsChecked — Function

ItemIsChecked(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Bool

Upstream link.

ImGuiTestEngine.ItemIsOpened — Function

ItemIsOpened(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Bool

Upstream link.

ImGuiTestEngine.ItemVerifyCheckedIfAlive — Function

ItemVerifyCheckedIfAlive(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    checked
)

Upstream link.

ImGuiTestEngine.ItemHold — Function

ItemHold(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    time
)

Upstream link.

ImGuiTestEngine.ItemHoldForFrames — Function

ItemHoldForFrames(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    frames
)

Upstream link.

ImGuiTestEngine.ItemDragOverAndHold — Function

ItemDragOverAndHold(
    ref_src::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    ref_dst::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.ItemDragAndDrop — Function

ItemDragAndDrop(
    ref_src::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    ref_dst::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
ItemDragAndDrop(
    ref_src::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    ref_dst::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    button
)

Upstream link.

ImGuiTestEngine.ItemDragWithDelta — Function

ItemDragWithDelta(
    ref_src::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    pos_delta::Union{CImGui.lib.ImVec2, Tuple{T, T} where T}
)

Upstream link.

ImGuiTestEngine.TabClose — Function

TabClose(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.TabBarCompareOrder — Function

TabBarCompareOrder(
    tab_bar::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiTabBar}},
    tab_order
) -> Bool

Upstream link.

ImGuiTestEngine.MenuAction — Function

MenuAction(
    action::ImGuiTestEngine.lib.ImGuiTestAction,
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.MenuActionAll — Function

MenuActionAll(
    action::ImGuiTestEngine.lib.ImGuiTestAction,
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.MenuClick — Function

MenuClick(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.MenuCheck — Function

MenuCheck(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.MenuUncheck — Function

MenuUncheck(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.MenuCheckAll — Function

MenuCheckAll(
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.MenuUncheckAll — Function

MenuUncheckAll(
    ref_parent::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.ComboClick — Function

ComboClick(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.ComboClickAll — Function

ComboClickAll(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)

Upstream link.

ImGuiTestEngine.TableOpenContextMenu — Function

TableOpenContextMenu(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
TableOpenContextMenu(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    column_n
)

Upstream link.

ImGuiTestEngine.TableClickHeader — Function

TableClickHeader(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    label
) -> CImGui.lib.ImGuiSortDirection
TableClickHeader(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    label,
    key_mods
) -> CImGui.lib.ImGuiSortDirection

Upstream link.

ImGuiTestEngine.TableSetColumnEnabled — Function

TableSetColumnEnabled(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    label,
    enabled
)

Upstream link.

ImGuiTestEngine.TableResizeColumn — Function

TableResizeColumn(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    column_n,
    width
)

Upstream link.

ImGuiTestEngine.TableGetSortSpecs — Function

TableGetSortSpecs(
    ref::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
) -> Ptr{CImGui.lib.ImGuiTableSortSpecs}

Upstream link.

ImGuiTestEngine.ViewportPlatform_SetWindowPos — Function

ViewportPlatform_SetWindowPos(
    viewport::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiViewport}},
    pos::Union{Ptr{Nothing}, Ref{CImGui.lib.ImVec2}, Ref{Tuple{T, T} where T}}
)

Upstream link.

ImGuiTestEngine.ViewportPlatform_SetWindowSize — Function

ViewportPlatform_SetWindowSize(
    viewport::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiViewport}},
    size::Union{Ptr{Nothing}, Ref{CImGui.lib.ImVec2}, Ref{Tuple{T, T} where T}}
)

Upstream link.

ImGuiTestEngine.ViewportPlatform_SetWindowFocus — Function

ViewportPlatform_SetWindowFocus(
    viewport::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiViewport}}
)

Upstream link.

ImGuiTestEngine.ViewportPlatform_CloseWindow — Function

ViewportPlatform_CloseWindow(
    viewport::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiViewport}}
)

Upstream link.

ImGuiTestEngine.DockClear — Function

DockClear(window_name)

Upstream link.

ImGuiTestEngine.DockInto — Function

DockInto(
    src_id::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    dst_id::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String}
)
DockInto(
    src_id::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    dst_id::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    split_dir::CImGui.lib.ImGuiDir
)
DockInto(
    src_id::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    dst_id::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    split_dir::CImGui.lib.ImGuiDir,
    is_outer_docking
)
DockInto(
    src_id::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    dst_id::Union{Int64, ImGuiTestEngine.lib.ImGuiTestRef, String},
    split_dir::CImGui.lib.ImGuiDir,
    is_outer_docking,
    flags
)

Upstream link.

ImGuiTestEngine.UndockNode — Function

UndockNode(dock_id)

Upstream link.

ImGuiTestEngine.UndockWindow — Function

UndockWindow(window_name)

Upstream link.

ImGuiTestEngine.WindowIsUndockedOrStandalone — Function

WindowIsUndockedOrStandalone(
    window::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiWindow}}
) -> Bool

Upstream link.

ImGuiTestEngine.DockIdIsUndockedOrStandalone — Function

DockIdIsUndockedOrStandalone(dock_id) -> Bool

Upstream link.

ImGuiTestEngine.DockNodeHideTabBar — Function

DockNodeHideTabBar(
    node::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiDockNode}},
    hidden
)

Upstream link.

ImGuiTestEngine.PerfCalcRef — Function

PerfCalcRef()

Upstream link.

ImGuiTestEngine.PerfCapture — Function

PerfCapture()
PerfCapture(category)
PerfCapture(category, test_name)
PerfCapture(category, test_name, csv_file)

Upstream link.

Other

These functions/types are less commonly used than the ones documented above.

ImGuiTestRef

ImGuiTestEngine.ImGuiTestRef — Type

ImGuiTestRef() -> ImGuiTestEngine.lib.ImGuiTestRef

Upstream link.

ImGuiTestRef(
    id::Integer
) -> ImGuiTestEngine.lib.ImGuiTestRef

Upstream link.

ImGuiTestRef(
    path::Union{Ptr{Int8}, Ptr{Nothing}, String}
) -> ImGuiTestEngine.lib.ImGuiTestRef

Upstream link.

ImGuiTestEngine.IsEmpty — Function

IsEmpty(self::Ptr{ImGuiTestEngine.lib.ImGuiTestLog}) -> Bool

Upstream link.

IsEmpty(self::Ptr{ImGuiTestEngine.lib.ImGuiTestRef}) -> Bool

Upstream link.

ImGuiTestRefDesc

ImGuiTestEngine.C_str — Function

C_str(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestRefDesc}
) -> Ptr{Int8}

Upstream link.

ImGuiTestEngine.ImGuiTestRefDesc — Type

ImGuiTestRefDesc(
    ref::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestRef}}
) -> ImGuiTestEngine.lib.ImGuiTestRefDesc

Upstream link.

ImGuiTestRefDesc(
    ref::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestRef}},
    item::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestItemInfo}}
) -> ImGuiTestEngine.lib.ImGuiTestRefDesc

Upstream link.

ImGuiTestGenericItemStatus

ImGuiTestEngine.ImGuiTestGenericItemStatus — Type

ImGuiTestGenericItemStatus(

) -> Ptr{ImGuiTestEngine.lib.ImGuiTestGenericItemStatus}

Upstream link.

ImGuiTestEngine.QuerySet — Function

QuerySet(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestGenericItemStatus}
)
QuerySet(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestGenericItemStatus},
    ret_val
)

Upstream link.

ImGuiTestEngine.QueryInc — Function

QueryInc(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestGenericItemStatus}
)
QueryInc(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestGenericItemStatus},
    ret_val
)

Upstream link.

ImGuiTestEngine.Draw — Function

Draw(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestGenericItemStatus}
)

Upstream link.

ImGuiTestItemList

ImGuiTestEngine.Reserve — Function

Reserve(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestItemList},
    capacity
)

Upstream link.

ImGuiTestEngine.GetSize — Function

GetSize(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestItemList}
) -> Int32

Upstream link.

ImGuiTestEngine.GetByIndex — Function

GetByIndex(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestItemList},
    n
) -> Ptr{ImGuiTestEngine.lib.ImGuiTestItemInfo}

Upstream link.

ImGuiTestEngine.GetByID — Function

GetByID(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestItemList},
    id
) -> Ptr{ImGuiTestEngine.lib.ImGuiTestItemInfo}

Upstream link.

ImGuiTestEngine.Size — Function

Size(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestItemList}
) -> UInt64

Upstream link.

ImGuiTestEngine.Begin — Function

Begin(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestItemList}
) -> Ptr{ImGuiTestEngine.lib.ImGuiTestItemInfo}

Upstream link.

ImGuiTestEngine.End — Function

End(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestItemList}
) -> Ptr{ImGuiTestEngine.lib.ImGuiTestItemInfo}

Upstream link.

ImGuiPerfTool

ImGuiTestEngine.ImGuiPerfTool — Type

ImGuiPerfTool() -> Ptr{ImGuiTestEngine.lib.ImGuiPerfTool}

Upstream link.

ImGuiTestEngine.Clear — Function

Clear(self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureImageBuf})

Free allocated memory buffer if such exists.

Upstream link.

Clear(self::Ptr{ImGuiTestEngine.lib.ImGuiTestItemList})

Upstream link.

Clear(self::Ptr{ImGuiTestEngine.lib.ImGuiTestLog})

Upstream link.

Clear(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestGenericItemStatus}
)

Upstream link.

Clear(self::Ptr{ImGuiTestEngine.lib.ImGuiTestGenericVars})

Upstream link.

Clear(self::Ptr{ImGuiTestEngine.lib.ImGuiCsvParser})

Free allocated buffers.

Upstream link.

Clear(self::Ptr{ImGuiTestEngine.lib.ImGuiTestGatherTask})

Warning

This function is internal, it may change in the future.

Upstream link.

Clear(self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool})

Upstream link.

ImGuiTestEngine.LoadCSV — Function

LoadCSV(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool}
) -> Bool
LoadCSV(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    filename
) -> Bool

Upstream link.

ImGuiTestEngine.AddEntry — Function

AddEntry(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    entry::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiPerfToolEntry}}
)

Upstream link.

ImGuiTestEngine.ShowPerfToolWindow — Function

ShowPerfToolWindow(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    engine::ImGuiTestEngine.Engine,
    p_open
)

Upstream link.

ImGuiTestEngine.ViewOnly — Function

ViewOnly(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    perf_name::Union{Ptr{Int8}, String}
)

Upstream link.

ViewOnly(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    perf_names::Ref{Vector{String}}
)

Upstream link.

ImGuiTestEngine.GetEntryByBatchIdx — Function

GetEntryByBatchIdx(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    idx
) -> Ptr{ImGuiTestEngine.lib.ImGuiPerfToolEntry}
GetEntryByBatchIdx(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    idx,
    perf_name
) -> Ptr{ImGuiTestEngine.lib.ImGuiPerfToolEntry}

Upstream link.

ImGuiTestEngine.SaveHtmlReport — Function

SaveHtmlReport(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    file_name
) -> Bool
SaveHtmlReport(
    self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool},
    file_name,
    image_file
) -> Bool

Upstream link.

ImGuiTestEngine.Empty — Function

Empty(self::Ptr{ImGuiTestEngine.lib.ImGuiPerfTool}) -> Bool

Upstream link.

ImGuiPerfToolEntry:

ImGuiTestEngine.ImGuiPerfToolEntry — Type

ImGuiPerfToolEntry(
    rhs::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiPerfToolEntry}}
) -> Ptr{ImGuiTestEngine.lib.ImGuiPerfToolEntry}

Upstream link.

ImGuiTestEngine.Set — Function

Upstream link.

ImGuiCaptureContext

ImGuiTestEngine.ImGuiCaptureContext — Type

ImGuiCaptureContext(

) -> Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
ImGuiCaptureContext(
    capture_func::Ref{Nothing}
) -> Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}

Upstream link.

ImGuiTestEngine.PreNewFrame — Function

PreNewFrame(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
)

Upstream link.

ImGuiTestEngine.PreRender — Function

PreRender(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
)

Upstream link.

ImGuiTestEngine.PostRender — Function

PostRender(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
)

Upstream link.

ImGuiTestEngine.CaptureUpdate — Function

CaptureUpdate(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext},
    args::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiCaptureArgs}}
) -> ImGuiTestEngine.lib.ImGuiCaptureStatus

Upstream link.

ImGuiTestEngine.RestoreBackedUpData — Function

RestoreBackedUpData(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
)

Upstream link.

ImGuiTestEngine.ClearState — Function

ClearState(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
)

Upstream link.

ImGuiTestEngine.BeginVideoCapture — Function

BeginVideoCapture(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext},
    args::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiCaptureArgs}}
)

Upstream link.

ImGuiTestEngine.EndVideoCapture — Function

EndVideoCapture(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
)

Upstream link.

ImGuiTestEngine.IsCapturingVideo — Function

IsCapturingVideo(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
) -> Bool

Upstream link.

ImGuiTestEngine.IsCapturing — Function

IsCapturing(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureContext}
) -> Bool

Upstream link.

ImGuiCaptureToolUI

ImGuiTestEngine.ImGuiCaptureToolUI — Type

ImGuiCaptureToolUI(

) -> Ptr{ImGuiTestEngine.lib.ImGuiCaptureToolUI}

Upstream link.

ImGuiTestEngine.ShowCaptureToolWindow — Function

ShowCaptureToolWindow(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureToolUI},
    context::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiCaptureContext}}
)
ShowCaptureToolWindow(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureToolUI},
    context::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiCaptureContext}},
    p_open
)

Render a capture tool window with various options and utilities.

Upstream link.

ImGuiCsvParser

ImGuiTestEngine.ImGuiCsvParser — Type

ImGuiCsvParser() -> Ptr{ImGuiTestEngine.lib.ImGuiCsvParser}
ImGuiCsvParser(
    columns
) -> Ptr{ImGuiTestEngine.lib.ImGuiCsvParser}

Upstream link.

ImGuiTestEngine.Load — Function

Load(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCsvParser},
    file_name
) -> Bool

Open and parse a CSV file.

Upstream link.

ImGuiTestEngine.GetCell — Function

GetCell(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCsvParser},
    row,
    col
) -> Ptr{Int8}

Upstream link.

ImGuiTestLog

ImGuiTestEngine.ImGuiTestLog — Type

ImGuiTestLog() -> Ptr{ImGuiTestEngine.lib.ImGuiTestLog}

Upstream link.

ImGuiTestEngine.ExtractLinesForVerboseLevels — Function

ExtractLinesForVerboseLevels(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestLog},
    level_min::ImGuiTestEngine.lib.ImGuiTestVerboseLevel,
    level_max::ImGuiTestEngine.lib.ImGuiTestVerboseLevel,
    out_buffer::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiTextBuffer}}
) -> Int32

Upstream link.

ImGuiTestEngine.UpdateLineOffsets — Function

UpdateLineOffsets(
    self::Ptr{ImGuiTestEngine.lib.ImGuiTestLog},
    engine_io::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiTestEngineIO}},
    level::ImGuiTestEngine.lib.ImGuiTestVerboseLevel,
    start
)

Upstream link.

ImGuiTestEngine.Destroy — Function

Destructor for ImGuiCaptureImageBuf

Destructor for ImGuiCaptureContext

Destructor for ImGuiCaptureToolUI

Destructor for ImGuiTestItemInfo

Destructor for ImGuiTestLog

Destructor for ImGuiTest

Destructor for ImGuiTestRef

Destructor for ImGuiTestRefDesc

Destructor for ImGuiTestActionFilter

Destructor for ImGuiTestGenericItemStatus

Destructor for ImGuiTestGenericVars

Destructor for ImGuiCsvParser

Destructor for ImGuiPerfToolEntry

Destructor for ImGuiPerfToolBatch

Destructor for ImGuiPerfTool

ImGuiTestEngine.ImHashDecoratedPath — Function

ImHashDecoratedPath(str) -> UInt32
ImHashDecoratedPath(str, str_end) -> UInt32
ImHashDecoratedPath(str, str_end, seed) -> UInt32

Upstream link.

ImGuiTestEngine.ImFindNextDecoratedPartInPath — Function

ImFindNextDecoratedPartInPath(str) -> Ptr{Int8}
ImFindNextDecoratedPartInPath(str, str_end) -> Ptr{Int8}

Upstream link.

ImGuiTestEngine.ImFileExist — Function

ImFileExist(filename) -> Bool

Upstream link.

ImGuiTestEngine.ImFileDelete — Function

ImFileDelete(filename) -> Bool

Upstream link.

ImGuiTestEngine.ImFileCreateDirectoryChain — Function

ImFileCreateDirectoryChain(path) -> Bool
ImFileCreateDirectoryChain(path, path_end) -> Bool

Upstream link.

ImGuiTestEngine.ImFileFindInParents — Function

ImFileFindInParents(
    sub_path,
    max_parent_count,
    output
) -> Bool

Upstream link.

ImGuiTestEngine.ImFileLoadSourceBlurb — Function

ImFileLoadSourceBlurb(
    filename,
    line_no_start,
    line_no_end,
    out_buf::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiTextBuffer}}
) -> Bool

Upstream link.

ImGuiTestEngine.ImGuiCaptureImageBuf — Type

ImGuiCaptureImageBuf(

) -> Ptr{ImGuiTestEngine.lib.ImGuiCaptureImageBuf}

Upstream link.

ImGuiTestEngine.ImPathFindFilename — Function

ImPathFindFilename(path) -> Ptr{Int8}
ImPathFindFilename(path, path_end) -> Ptr{Int8}

Return value always between path and path_end.

Upstream link.

ImGuiTestEngine.ImPathFindExtension — Function

ImPathFindExtension(path) -> Ptr{Int8}
ImPathFindExtension(path, path_end) -> Ptr{Int8}

Return value always between path and path_end.

Upstream link.

ImGuiTestEngine.ImPathFixSeparatorsForCurrentOS — Function

ImPathFixSeparatorsForCurrentOS(buf)

Upstream link.

ImGuiTestEngine.CreateEmpty — Function

CreateEmpty(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureImageBuf},
    w,
    h
)

Reallocate buffer for pixel data and zero it.

Upstream link.

ImGuiTestEngine.SaveFile — Function

SaveFile(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureImageBuf},
    filename
) -> Bool

Save pixel data to specified image file.

Upstream link.

ImGuiTestEngine.RemoveAlpha — Function

RemoveAlpha(
    self::Ptr{ImGuiTestEngine.lib.ImGuiCaptureImageBuf}
)

Clear alpha channel from all pixels.

Upstream link.

ImGuiTestEngine.ImStrReplace — Function

ImStrReplace(s, find, repl)

Upstream link.

ImGuiTestEngine.ImStrchrRangeWithEscaping — Function

ImStrchrRangeWithEscaping(str, str_end, find_c) -> Ptr{Int8}

Upstream link.

ImGuiTestEngine.ImStrXmlEscape — Function

ImStrXmlEscape(s)

Upstream link.

ImGuiTestEngine.ImStrBase64Encode — Function

ImStrBase64Encode(src, dst, length) -> Int32

Upstream link.

ImGuiTestEngine.ImParseExtractArgcArgvFromCommandLine — Function

ImParseExtractArgcArgvFromCommandLine(
    out_argc,
    out_argv,
    cmd_line
)

Upstream link.

ImGuiTestEngine.ImParseFindIniSection — Function

ImParseFindIniSection(
    ini_config,
    header,
    result::Union{Ptr{Nothing}, Ref{CImGui.lib.ImVector_char}}
) -> Bool

Upstream link.

ImGuiTestEngine.ImTimeGetInMicroseconds — Function

ImTimeGetInMicroseconds() -> UInt64

Upstream link.

ImGuiTestEngine.ImTimestampToISO8601 — Function

ImTimestampToISO8601(timestamp, out_date)

Upstream link.

ImGuiTestEngine.ImThreadSleepInMilliseconds — Function

ImThreadSleepInMilliseconds(ms)

Upstream link.

ImGuiTestEngine.ImThreadSetCurrentThreadDescription — Function

ImThreadSetCurrentThreadDescription(description)

Upstream link.

ImGuiTestEngine.ImBuildGetCompilationInfo — Function

ImBuildGetCompilationInfo(

) -> Ptr{ImGuiTestEngine.lib.ImBuildInfo}

Upstream link.

ImGuiTestEngine.ImBuildFindGitBranchName — Function

ImBuildFindGitBranchName(git_repo_path, branch_name) -> Bool

Upstream link.

ImGuiTestEngine.ImOsCreateProcess — Function

ImOsCreateProcess(cmd_line) -> Bool

Upstream link.

ImGuiTestEngine.ImOsPOpen — Function

ImOsPOpen(cmd_line, mode) -> Ptr{Base.Libc.FILE}

Upstream link.

ImGuiTestEngine.ImOsPClose — Function

ImOsPClose(fp)

Upstream link.

ImGuiTestEngine.ImOsOpenInShell — Function

ImOsOpenInShell(path)

Upstream link.

ImGuiTestEngine.ImOsIsDebuggerPresent — Function

ImOsIsDebuggerPresent() -> Bool

Upstream link.

ImGuiTestEngine.ImOsOutputDebugString — Function

ImOsOutputDebugString(message)

Upstream link.

ImGuiTestEngine.ImOsConsoleSetTextColor — Function

ImOsConsoleSetTextColor(
    stream::ImGuiTestEngine.lib.ImOsConsoleStream,
    color::ImGuiTestEngine.lib.ImOsConsoleTextColor
)

Upstream link.

ImGuiTestEngine.TableGetHeaderID — Function

TableGetHeaderID(
    table::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiTable}},
    column::Union{Ptr{Int8}, Ptr{Nothing}, String}
) -> UInt32
TableGetHeaderID(
    table::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiTable}},
    column::Union{Ptr{Int8}, Ptr{Nothing}, String},
    instance_no::Integer
) -> UInt32

Upstream link.

TableGetHeaderID(
    table::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiTable}},
    column_n::Integer
) -> UInt32
TableGetHeaderID(
    table::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiTable}},
    column_n::Integer,
    instance_no::Integer
) -> UInt32

Upstream link.

ImGuiTestEngine.TableDiscardInstanceAndSettings — Function

TableDiscardInstanceAndSettings(table_id)

Upstream link.

ImGuiTestEngine.DrawDataVerifyMatchingBufferCount — Function

DrawDataVerifyMatchingBufferCount(
    draw_data::Union{Ptr{Nothing}, Ref{CImGui.lib.ImDrawData}}
)

Upstream link.

ImGuiTestEngine.PerfToolAppendToCSV — Function

PerfToolAppendToCSV(
    perf_log::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiPerfTool}},
    entry::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiPerfToolEntry}}
)
PerfToolAppendToCSV(
    perf_log::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiPerfTool}},
    entry::Union{Ptr{Nothing}, Ref{ImGuiTestEngine.lib.ImGuiPerfToolEntry}},
    filename
)

Upstream link.

ImGuiTestEngine.ImGuiTestActionFilter — Type

ImGuiTestActionFilter(

) -> Ptr{ImGuiTestEngine.lib.ImGuiTestActionFilter}

Upstream link.

ImGuiTestEngine.ImGuiTestEngineHook_ItemAdd — Function

ImGuiTestEngineHook_ItemAdd(
    ui_ctx::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiContext}},
    id,
    bb::Union{Ptr{Nothing}, Ref{CImGui.lib.ImRect}},
    item_data::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiLastItemData}}
)

Upstream link.

ImGuiTestEngine.ImGuiTestEngineHook_ItemInfo — Function

ImGuiTestEngineHook_ItemInfo(
    ui_ctx::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiContext}},
    id,
    label,
    flags
)

Upstream link.

ImGuiTestEngine.ImGuiTestEngineHook_Log — Function

ImGuiTestEngineHook_Log(
    ui_ctx::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiContext}},
    fmt
)

Upstream link.

ImGuiTestEngine.FindItemDebugLabel — Function

FindItemDebugLabel(
    ui_ctx::Union{Ptr{Nothing}, Ref{CImGui.lib.ImGuiContext}},
    id
) -> Ptr{Int8}

Upstream link.

ImGuiTestEngine.Check — Function

Check(file, func, line, flags, result, expr) -> Bool

Upstream link.

ImGuiTestEngine.CheckStrOp — Function

CheckStrOp(
    file,
    func,
    line,
    flags,
    op,
    lhs_var,
    lhs_value,
    rhs_var,
    rhs_value,
    out_result
) -> Bool

Upstream link.

ImGuiTestEngine.Error — Function

Error(file, func, line, flags, fmt) -> Bool

Upstream link.

ImGuiTestEngine.AssertLog — Function

AssertLog(expr, file, _function, line)

Upstream link.

ImGuiTestEngine.GetTempStringBuilder — Function

GetTempStringBuilder() -> Ptr{CImGui.lib.ImGuiTextBuffer}

Upstream link.

ImGuiTestEngine.ImGuiTestGenericVars — Type

ImGuiTestGenericVars(

) -> Ptr{ImGuiTestEngine.lib.ImGuiTestGenericVars}

Upstream link.

ImGuiTestEngine.InstallDefaultCrashHandler — Function

InstallDefaultCrashHandler()

Install default crash handler (if you don't have one).

Upstream link.

ImGuiTestEngine.CrashHandler — Function

CrashHandler()

Default crash handler, should be called from a custom crash handler if such exists.

Upstream link.

ImGuiTestEngine.ImGuiTestItemInfo — Type

ImGuiTestItemInfo(

) -> Ptr{ImGuiTestEngine.lib.ImGuiTestItemInfo}

Upstream link.